# gh-121192: Add fast path to `deepcopy()` for empty list/tuple/dict/set