--- a/Misc/NEWS.d/next/Library/2024-07-01-00-32-33.gh-issue-121192.RZFhQP.rst
+++ b/Misc/NEWS.d/next/Library/2024-07-01-00-32-33.gh-issue-121192.RZFhQP.rst
@@ -1,0 +1,2 @@
+Add fast path to :func:`copy.deepcopy` for empty lists, tuples, dicts, sets
+or frozensets.