`deepcopy()` can be surprisingly slow when called with empty containers like lists, tuples, dicts, sets or frozensets.

This adds a fast path for this case similar to #114266 which speeds up these cases by about 4x - 20x while having a small impact on the default path.

Here are some benchmarks comparing this PR against `main`. Let me know if there are more benchmarks that I can run to verify that there a no performance regressions:
```python
import pyperf
runner = pyperf.Runner()

setup = """
import copy

a = {"list": [1, 2 ,3 ,4], "t": (1, 2, 3), "str": "hello", "subdict": {"a": True}}

t = ()
fs = frozenset()
l = []
s = set()
d = {}
deep = [[], (), {}, set(), frozenset()]
"""

runner.timeit(name="deepcopy dict", stmt=f"b = copy.deepcopy(a)", setup=setup)
runner.timeit(name="deepcopy empty tuple", stmt=f"b = copy.deepcopy(t)", setup=setup)
runner.timeit(name="deepcopy empty frozenset", stmt=f"b = copy.deepcopy(fs)", setup=setup)
runner.timeit(name="deepcopy empty list", stmt=f"b = copy.deepcopy(l)", setup=setup)
runner.timeit(name="deepcopy empty set", stmt=f"b = copy.deepcopy(s)", setup=setup)
runner.timeit(name="deepcopy empty dict", stmt=f"b = copy.deepcopy(d)", setup=setup)
runner.timeit(name="deepcopy multiple empty containers", stmt=f"b = copy.deepcopy(deep)", setup=setup)

```
```
deepcopy dict: Mean +- std dev: [baseline] 1.86 us +- 0.06 us -> [less-optimize-empty-copy] 1.97 us +- 0.02 us: 1.06x slower
deepcopy empty tuple: Mean +- std dev: [baseline] 285 ns +- 2 ns -> [less-optimize-empty-copy] 55.3 ns +- 1.6 ns: 5.15x faster
deepcopy empty frozenset: Mean +- std dev: [baseline] 1.47 us +- 0.11 us -> [less-optimize-empty-copy] 73.7 ns +- 4.7 ns: 19.96x faster
deepcopy empty list: Mean +- std dev: [baseline] 323 ns +- 2 ns -> [less-optimize-empty-copy] 74.1 ns +- 3.0 ns: 4.36x faster
deepcopy empty set: Mean +- std dev: [baseline] 1.46 us +- 0.10 us -> [less-optimize-empty-copy] 77.4 ns +- 2.3 ns: 18.80x faster
deepcopy empty dict: Mean +- std dev: [baseline] 326 ns +- 4 ns -> [less-optimize-empty-copy] 75.3 ns +- 3.0 ns: 4.33x faster
deepcopy multiple empty containers: Mean +- std dev: [baseline] 4.13 us +- 0.04 us -> [less-optimize-empty-copy] 1.33 us +- 0.09 us: 3.10x faster

Geometric mean: 5.22x faster
```