@lgeiger 2024-07-01
I ran the relevant pyperformance deepcopy benchmark which does seem to suggest that this change makes the non empty case slightly slower:
```
Performance version: 1.11.0
Python version: 3.14.0a0 (64-bit) revision 1a84bdc237
Report on macOS-14.5-arm64-arm-64bit-Mach-O
Number of logical CPUs: 8
Start date: 2024-07-01 02:40:30.390645
End date: 2024-07-01 02:41:08.674496

### deepcopy ###
Mean +- std dev: 118 us +- 2 us -> 124 us +- 2 us: 1.05x slower
Significant (t=-16.39)

### deepcopy_memo ###
Mean +- std dev: 14.2 us +- 0.4 us -> 14.6 us +- 1.3 us: 1.03x slower
Significant (t=-2.32)

### deepcopy_reduce ###
Mean +- std dev: 1.21 us +- 0.02 us -> 1.25 us +- 0.02 us: 1.03x slower
Significant (t=-11.48
```