@eendebakpt 2024-07-02
The optimization applied in the latest iteration of the PR is only applied when no memo is passed to `deepcopy`. The advantage is that any recursive structures like `{'a': [1, 2, 3, 4, ,5]}` will have no (or near zero) overhead. So I agree that performance regression for many common cases is not a big concern. 

The question then is whether a `deepcopy` directly on an empty builtin like like and set (and not a `deepcopy` on a recursive structure containing empty builtins, e.g, `( {}, [], ())`) is a common enough operation to apply this PR. I am not convinced on this, In the corresponding issue an example is given for pydantic basemodels. But there the deepcopy overhead can be avoided by using a default factory. Also the corresponding situation for plain dataclasses cannot occur:
```
from dataclasses import dataclass

@dataclass
class DC:
    a : list = []
    
x=DC()
```
raises a `ValueError: mutable default <class 'list'> for field a is not allowed: use default_factory`.  With `attrs` an empty list is allowed, but no deepcopy is performed on instantiation. For example
```
import attrs
@attrs.define
class DC:
    a : list = []
    
x=DC()
y=DC()

x.a.append(1)

print(x)
print(y)
```
results in 
```
DC(a=[1])
DC(a=[1])
```