@sobolevn 2024-07-03
@eendebakpt it can happen for dataclass, example:

```python
from dataclasses import dataclass, as_dict

@dataclass 
class Some:
    field: set[int]

as_dict(Some(set()))
```