@eendebakpt 2024-07-03
> @eendebakpt it can happen for dataclass, example:

Indeed. Here a more complete example, including the cases where this PR helps
```
from dataclasses import dataclass, asdict

@dataclass 
class Some:
    a: str
    t: tuple[int]
    field: set[int] # benefits
    l : list[int]
    d : dict[str, str]
    fs: frozenset[int] # benefits
    nonempty : set[int]

s=Some('x', tuple(), set(), [], d={}, fs=frozenset(), nonempty={1, 2})
asdict(s)
```