@sobolevn 2024-07-03
Other use-cases that we can find in our own code:
- dataclass: https://github.com/python/cpython/blob/26d24eeb90d781e381b97d64b4dcb1ee4dd891fe/Lib/dataclasses.py#L1487 also `as_tuple` for `set` and `frozenset`
- weakref: `WeakValueDictionary` https://github.com/python/cpython/blob/26d24eeb90d781e381b97d64b4dcb1ee4dd891fe/Lib/weakref.py#L191 and `WeakKeyDictionary`: https://github.com/python/cpython/blob/26d24eeb90d781e381b97d64b4dcb1ee4dd891fe/Lib/weakref.py#L448 where values can be any of the given types, and keys can be empty tuples and empty frozendicts (which is rather rare, I agree)
- http: https://github.com/python/cpython/blob/26d24eeb90d781e381b97d64b4dcb1ee4dd891fe/Lib/http/cookiejar.py#L1819 where `self._cookies` is empty by default https://github.com/python/cpython/blob/26d24eeb90d781e381b97d64b4dcb1ee4dd891fe/Lib/http/cookiejar.py#L1267
- coping nested structures where keys/values are empty collections, like: `{'data': []}`